Fix selection of an answer to a question
Had to pass a whole new data structure to setQuizData, rather than a tweaked version, because React doesn't take a deep look at the data, so was not updating the UI to show selected state.

Also had to pass in the question ID so that selecting one answer didn't unselect all of the others.