Support restarting the game
Made fetchData() resuable - because it's needed both at the start of the game, to load the game data, and on restart. Put it further down in the app on the assumption that it's better to have it placed after definition/declaration of the other functions/variables it makes use of.